show executions close #4

+++ actions/tasks.rb
@@ -5,6 +5,15 @@ class HTTPCronApi < Sinatra::Base
     current_user.tasks.to_json
   end
 
+  get '/tasks/:id/executions' do |id|
+    t = Task.find(id)
+    unless t.user.admin || t.user == current_user
+      halt 403, 'This Task do not belongs to you!'
+    end
+    content_type :json
+    t.executions.to_json
+  end
+
   post '/tasks' do
     check_parameter_for_blank :name, :url, :cron
 
@@ -51,6 +60,9 @@ class HTTPCronApi < Sinatra::Base
     end
 
     t = Task.find(id)
+    unless t.user.admin || t.user == current_user
+      halt 403, 'This Task do not belongs to you!'
+    end
     t[:name] = params[:name]
     t[:url] = params[:url]
     t[:cron] = params[:cron]
@@ -73,6 +85,9 @@ class HTTPCronApi < Sinatra::Base
 
   delete '/tasks/:id' do |id|
     t = Task.find(id)
+    unless t.user.admin || t.user == current_user
+      halt 403, 'This Task do not belongs to you!'
+    end
     begin
       t.destroy
     rescue Exception => e
